refactor: hooks useLogin/useHomeRedirect, types UseDishOptions/UseDeleteDishOptions, shared components, AppRoutes, tags/dishes
