+++ frontend/src/components/AppRoutes.tsx
@@ -1,7 +1,7 @@
 import { Routes, Route, Navigate } from 'react-router-dom'
-import Login from '../shared/pages/Login'
-import Register from '../shared/pages/Register'
-import Home from '../shared/pages/Home'
+import Login from '../shared/components/Login'
+import Register from '../shared/components/Register'
+import Home from '../shared/components/Home'
 import Menu from '../client/pages/Menu'
 import RestaurantsList from '../client/pages/RestaurantsList'
 import CreateEstablishment from '../owner/components/Establishment/CreateEstablishment'
