+++ frontend/src/css/owner/Dishes.css
@@ -27,46 +27,6 @@
   color: var(--text-primary);
 }
 
-.dishes-search-container {
-  margin-bottom: var(--spacing-lg);
-}
-
-.dishes-search-wrapper {
-  position: relative;
-  max-width: 400px;
-}
-
-.dishes-search-icon {
-  position: absolute;
-  left: var(--spacing-md);
-  top: 50%;
-  transform: translateY(-50%);
-  font-size: var(--font-size-lg);
-  color: var(--text-secondary);
-  pointer-events: none;
-}
-
-.dishes-search-input {
-  width: 100%;
-  padding: var(--spacing-md) var(--spacing-md) var(--spacing-md) 48px;
-  background: var(--color-bg-secondary);
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  border-radius: var(--border-radius-lg);
-  color: var(--text-primary);
-  font-size: var(--font-size-base);
-  transition: all var(--transition-base);
-}
-
-.dishes-search-input::placeholder {
-  color: var(--text-secondary);
-}
-
-.dishes-search-input:focus {
-  outline: none;
-  border-color: var(--color-primary);
-  box-shadow: 0 0 0 3px rgba(255, 127, 63, 0.1);
-}
-
 .dishes-categories {
   display: flex;
   gap: var(--spacing-md);
