+++ frontend/src/owner/hooks/errorHandler.ts
@@ -1,6 +1,4 @@
-/**
- * Trata erros da API que podem vir como string ou array de strings
- */
+// Trata erros da API que podem vir como string ou array de strings
 export function formatApiError(error: string | string[] | undefined): string {
   if (!error) return ''
   
@@ -11,9 +9,7 @@ export function formatApiError(error: string | string[] | undefined): string {
   return error
 }
 
-/**
- * Extrai mensagem de erro de uma resposta da API
- */
+//retorna a mensagem de erro da resposta da api
 export function getErrorMessage(response: { error?: string | string[], errors?: string[] }): string {
   if (response.errors && Array.isArray(response.errors) && response.errors.length > 0) {
     return formatApiError(response.errors)
