+++ frontend/src/shared/components/Home.tsx
@@ -1,0 +1,11 @@
+import { useHomeRedirect } from '../hooks/useHomeRedirect'
+
+export default function Home() {
+  useHomeRedirect()
+
+  return (
+    <div style={{ padding: '20px', textAlign: 'center' }}>
+      <p>Redirecionando...</p>
+    </div>
+  )
+}
