+++ frontend/src/shared/components/Login.tsx
@@ -1,0 +1,79 @@
+import { Link } from 'react-router-dom'
+import { useLogin } from '../hooks/useLogin'
+import '../../css/shared/Login.css'
+
+export default function Login() {
+  const {
+    email,
+    setEmail,
+    password,
+    setPassword,
+    error,
+    successMessage,
+    loading,
+    handleSubmit,
+  } = useLogin()
+
+  return (
+    <div className="login-container">
+      <div className="login-card">
+        <h1>PÁLEVÁ</h1>
+        <h2>Login</h2>
+
+        <form onSubmit={handleSubmit}>
+          <div className="form-group">
+            <label htmlFor="email">Email</label>
+            <input
+              id="email"
+              type="email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              placeholder="seu@email.com"
+              required
+              disabled={loading}
+            />
+          </div>
+
+          <div className="form-group">
+            <label htmlFor="password">Senha</label>
+            <input
+              id="password"
+              type="password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              placeholder="••••••••"
+              required
+              disabled={loading}
+            />
+          </div>
+
+          {successMessage && (
+            <div className="success-message">
+              {successMessage}
+            </div>
+          )}
+
+          {error && (
+            <div className="error-message">
+              {error}
+            </div>
+          )}
+
+          <button
+            type="submit"
+            className="btn-primary"
+            disabled={loading}
+          >
+            {loading ? 'Entrando...' : 'Entrar'}
+          </button>
+
+          <div className="login-footer">
+            <p>
+              Não tem uma conta? <Link to="/register">Cadastre-se</Link>
+            </p>
+          </div>
+        </form>
+      </div>
+    </div>
+  )
+}
